backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch>=2.2.0
torchvision>=0.17.0
Pillow==10.0.1
python-dotenv==1.0.0 